Move constants from config to _contents
