+++ apps/nextjs-website/src/_contents/auth.ts
@@ -1,0 +1,15 @@
+export const companyRoles = [
+  'ente-pubblico',
+  'partner-tecnologico',
+  'psp',
+  'gestore-di-pubblico-servizio',
+  'azienda-privata',
+  'altro',
+];
+
+export const signUpAdvantages = [
+  'exclusive_contents',
+  'product_updates',
+  'api_keys',
+  'support',
+];
